requirements.txt
pyqt6
pillow
opencv-python
transformers
torch